Add acceptance scenario for worker class that moves file to public ftp
To complete the acceptance scenario that test that the curator can trigger
the job to move files of dataset from the drop box to the public ftp,
I've added the acceptance scenario that test that the triggered job results
in files actually being physically copied.

We also need to configure root path and path alias correctly in console/config/main.php
for the copy to actually happen.

The implementation of acceptance scenario steps required:
* connecting to the queue monitor web app so we can check that the job has been completed
 * the host name needed to be added to the console and chrome container service configuration in docker-compose.yml
* ensuring that the queue is cleared after each test run, so we add a call to monitor/clear-all console commands in
the _after hook function of FilesystemCleanupHook.php Codeception module
(may not be the best location as it's not directly related to filesystem. I will think of better location)
* we need to physically create the files in the steps that generate examples uploads in ReviewerSteps.php, so the file move job
actually has files to copy
* and therefore in FilesystemCleanupHook.php, we need to add the destination directory (/var/ftp/public) to the list of directories
to clear out after the test suite has run