Add docker container service for queue workers
As we have File Upload Wizard API posting job in the Beanstalkd message queue,
we need a new container service that continually listen to the queue for new jobs
and then execute the workers job classes.

We use systemd for init system as is implement the Linux API, is used in more recent Debian distribution
and is very powerful. yii2-queue docs explain how to configure a listener with systemd but before that
setting up a non-privileged container with systemd is quite involved.
So we have a special Dockerfile (fuw/app/backend/Worker-Dockerfile) that bring instructions for setting up
a Yii2 application (as the workers are part of Yii2) from the regular Dockerfile for Yii2 apps (fuw/app/common/Dockerfile)
and add the systemd specificities I've copied from (alongside ancillary files):
https://github.com/dramaturg/docker-debian-systemd/blob/master/Dockerfile

I've added a docker-compose.yml service (fuw-worker) that use the aforementioned Worker-Dockerfile and we add the options to
support systemd idiosyncrasies, learned from:
https://developers.redhat.com/blog/2016/09/13/running-systemd-in-a-non-privileged-container/

the container doesn't require a php-fpm daemon as systemd daemon loaded as PID1 take cares of maintaining the listener up.
Therefore we choose for base image "php-7.2-cli-stretch" instead of "php-7.2-fpm-stretch"

We couldn't reuse the regular Dockerfile and add conditional directive (INSTALL_*) for systemd like we do for other functions
because setting systemd require more than installing binaries and copying files that we have to use other Docker directives
that cannot be contained in a conditional bash block, and it doesn't use php-fpm.

The Worker-Dockerfile is in fuw/app/backend rather than the fuw/app/common like the regular Dockerfile because message queues and their workers
are always backoffice systems.

It's not fully satisfying configuration yet, as I've not figured out yet how to get the logging out of that container.

This commit will also fix the current acceptance tests failure on CI because there's no queue listener service wasn't existing yet
(I used to run the queue listener directly in dev environment) hence the change to .gitlab-ci.yml to start up the fuw-worker container service
and its beantalkd dependent service

Finally, we add build instructions for fuw-worker in .gitlab-ci.yml so it benefits of the caching of container images