Docs(/site/help): Update text in API tab on help page
Updated text in Application programming interface tab on GigaDB website /site/help page as requested by @tuli for issue #198.

@tuli requested changes for the following text at http://gigadb.org/site/help#0.1_API:

The current API version is available on our main production database. This version will be periodically updated with new additional functionality, we will whenever possible maintain backwards compatability, but occassionally this may not be possible, for this reason we recomend regularly checking and updating you usage of our API.

To be changed to:

The current API version is available on our main production database. This version will be periodically updated with new additional functionality and we will whenever possible maintain backwards compatibility. Occasionally this may not be possible and for this reason we recommend regularly checking and updating your usage of our API.

The changes in text to the help.html page was implemented in two places:

* protected/views/site/help.php. Changes in this file results in visible changes in the help.html page
* ops/configuration/yii-config/help.html.dist. The same text was updated in this file for consistency reasons.

We need to decide whether to keep these *.dist files in the future.

List of work done:

* Change API description in help page

* To remove space at the end

* update help API text

* Update help API text

* Revert to commit:3954c0b0

* Update help API text in help.html.dist

* Update help API text in help.php

* Remove .help.php.swp