Change how adding metadata for uploaded file is triggered and done
Authors can upload files through the wizard or by ftp. Then we need to the attached metadata to FUW.
Because we cannot put any hook to ftp, we have a watcher service to monitor the filesystem and trigger metadata insert.
That method cause latency and flimsiness due to large number of filesystem events and race conditions.

We keep the watcher service for ftp upload, but for files uploaded through the wizard (which is the official and normal way),
we had a post-finish hook (in fuw/tusd/hooks/post-finish) to the Tusd upload service that will call a Yii2 console commands in File Upload Wizard.

That entails a new console controller and a business object model, TusdController.php and Tusd.php respectively.

Because FUW's console application didn't have any code, Codeception testing is not initialized, so we need to set it up
the main configs being codeception.yml, unit.suite.yml and functional.suite.yml.
After that, using ``codecept bootstrap --namespace 'console\tests'`` command,
Codeception generates all the supporting code necessary to run the tests.
We then add the fixture data necessary for testing adding metadata upon post-finish tusd hook trigger (in console/tests/_data).
The unit test and functional tests are TusdTest.php and TusdCest.php respectively.
Because the metadata saving flow is now more reactive, we can remove the 30/60 seconds wait step in acceptance scenarios, so they can run faster.

We changed the configuration of the watcher service to stop triggering database
update script (file-to-db.php) when file are moved to drop box filesystem after being uploaded through the wizard
That's what the removal of conf/files-metadata-to-database.conf is for.
(ftp uploads still triggers the database update script)