Use Flysystem in console command that processes ftp uploads
Flysystem allow us to abstract the filesystem so we can change the type of filestorage
without changing the code. This is important because we are still figuring out which cloud platform
gigadb and FUW will be deployed to.

The subsystem that move files to public ftp already uses Flysystem. This commit is for the FtpController.php.
Next will do the same for the console command to process wizard file upload through TUS.

The main changes are to replace absolute file path with path relative to the Flysystem's configured root path
(which differs whether were are on test environment or not, hence the change to test.php)
in code that deal with file operations.

I also needed to homogenize the bind mounted path in all container services that need to execute the console commands
(hence the change to docker-compose.yml to update the watcher container service bind mount paths and to watcher config files and scripts)