Update configuration for terraform and ansible to support new version
newer version of terraform and ansible have introduced changes that breaks the current configuration.

remove terraform-inventory binary from codebase as it can be installed separately and prevent keeping out-of-date versions.
In its place we have a bash script called terraform-inventory.sh that will call the separately installed binary.
For terraform-inventory to function and prevent the error

```
Error reading tfstate file: 0.12 format error: <nil>; pre-0.12 format error
```

the terraform environment variable TF_STATE needs to be set as a relative path to the .tfstate file

Updated the deployment documentation to mention the relevant changes