Fix bug in function captchaGenerator in SiteController
The captchaGenerator function used imagejpeg() to create an image
with a hardcoded `.png` file extension. This function has been
updated to use imagepng() to create an actual PNG image file.

CaptchaImageTest has been updated to test for a png image file,
not just an image.

Thanks to @rija for the above suggestions.