Fix: Website images display in local and staging GigaDB
In a local GigaDB website, the images directory has been mounted 
into the web container so new image files in it become immediately 
available to the website. Previously, files in the images directory 
were copied into a layer via Web-Dockerfile so new images added 
afterwards into the directory were not available when you link to 
them from HTML files.

A new Production-Web-Dockerfile reinstating the COPY 
`images /var/www/images` instruction is used for the web 
container for staging GigaDB so that it again has access to website 
image files.

Captcha images are not being displayed on a staged GigaDB. These 
captcha images are programmatically created by the production_app 
container in /var/www/images/tempcaptcha which is not available to 
the web container which is why they were not being seen on the website. 
This directory is now mounted as a volume and made available to the 
web container via docker-compose.staging.yml.

CaptchaImageTest.php provides a functional test to check for the 
presence of a captcha image on the contact web page at /site/contact. 
It uses a new function called getTagsWithSessionAndUrl($url, $tag) in 
BrowserPageSteps.php to return HTML img tags in the /site/contact 
web page. The functional test looks for the captcha img tag to parse
out its src metadata which it uses to reconstruct the URL for the 
captcha image. This image is then downloaded and checked that it is 
a png image with the expected size dimensions.