Add top operation stage and configure target environment from gitlab
To help debugging and operation on cloud deployment, added a manual stage to list all processes
running on all container.

In order for the pipeline to be usable on cloud environments, not just staging,
we need to configure GIGADB_ENV value from GITLAB environment variables specifics to a pipeline project.
(i.e: branch)

And we need to pull the right certificate files for the right environment (staging or live)