Enable Docker daemon to listen from multiple sockets SSL and non-SSL
We use Docker-PHP for inter-container communication, however it uses PHP's stream_socket_client()
which doesn't like the format of SSL certificates created by Ansible' role for making docker secure.
It is unclear if it's the version of PHP or that of the underlying OpenSSL library, but there are more correct ways
to fix this.

Docker daemon was configured according to Docker documentation for supporting
mutliple listening endpoints for connections by passing the daemon multiple -H parameters

However that would not work as soon as the daemon is expected to work
with ssl (by passing the --tslverify and related parameters) even if we pass the appropriate -H parameters
as whatever configured connection we go through it passes through TLS endpoint first
(and make PHP to trigger the error "Client sent an HTTP request to an HTTPS server")

There is undocumented solution (using systemd socket) from Docker (more details here: https://stackoverflow.com/a/44653660),
but this is to allow connecting locally to the daemon use Docker unix socket.

But we don't want to go back to the unsecure way of  mounting docker unix socket into containers, so instead we adapt the solution to
allow a local connection to TCP port 2375