Restrict package drewm/mailchimp to v2.5 only
The latest version (2.5.4) of the drewm/mailchimp-api package was being
downloaded and used which breaks NewsletterServiceTest. Version 2.5.4 is
downloaded because composer.json.dist specifies it as ^v2.5. Version
2.5.4 includes changing subscriberHash to a static function which might
be the cause of the test breaking. If "drewm/mailchimp-api": "v2.5"
is downloaded then this allows NewsletterServiceTest to pass.

This commit relates to gigascience/gigadb-website#410