Tests: Fix NewsletterServiceTest
The latest version (2.5.4) of the drewm/mailchimp-api package was being
downloaded and used which breaks NewsletterServiceTest. Version 2.5.4 is
downloaded because composer.json.dist specifies it as ^v2.5. Version
2.5.4 includes changing subscriberHash to a static function which might
be the cause of the test breaking. If "drewm/mailchimp-api": "v2.5.3"
is downloaded then this is the most recent version that allows 
NewsletterServiceTest to pass.