Remove previous socat container before starting it again
If socat container exists but is stopped, the manual gitlab stage to kick it off won't work
becausethe stopped container is still registered, so we need to remove any previous socat
container beforehand. We need to suffix the step with ``|| true`` so we can  ignore the error
when there was no previous socat container.