Fix 2 bugs when adding references to file samples in File Upload Wizard
As with the attributes fixed in the previous commit, we need to fix the life cycle
of the Sampler Vue.js component, so that the current instance is destroyed when the drawer panel for adding samples
in the medadata form of the wizard  is closed.

Secondly, the confirmation message after the author save and complete the wizard flow was misleading
it was showing sample update count (of 1) for uploads for which no samples was added.

Also replaced the upload id in the message with the clearer upload name.

Update a functional test and an acceptance test to cover these changes.