Fix Pager component to stop showing up when there's no file uploads
It is meant that the Next button only shows when the file uploads complete
or that there are files that have been previously uploaded.
A bug in the condition for testing whether to show the button made it so that
the Next button was showing when no uploads existed.
Changed the automated tests to reproduce the issue and fixed the problem.