Fix: Left-align placeholder text in front search bar
The import of destyle.less has been commented out in current.less. This allows the placeholder text to be left-aligned again in the front search bar. Having CSS reset functionality seems to be a good idea so that styling of all HTML elements are reset to a baseline that is consistent for all browsers and avoids cross-browser differences. For this reason, import of destyle.less has been commented out for the time being till when we can take another look at the CSS problem with GigaDB.

Relates to gigascience/gigadb-website#522.