Enable file upload button with DataPending status for a dataset
Authors normally upload files when the dataset status is "UserUploadingData".

However the curators have the ability to set the status to "DataPending"
during their review of upload files to indicate that files or data is missing.

Therefore the author should be able to access and operate the file upload wizard
with that status as well.

I have added a new acceptance test scenario to test that new valid status for file uploads.