Add Article container example to the style guide
In File Upload Wizard, the forms are wrapped by an HTML5 <article> element.
That needs to be styled so its elements do not stick to the bottom edge.
I've added a layout element to the style guide to illustrate that.

Also fixed viewing the style guide in the app by mounting its directory into the application container.
(as we no longer blanket mount the whole project directory for efficiency)