Fix catch-22 issue between backoffice workers and functional tests
DB migrations need to be run for gigadb-worker to start properly, but the migrations
were run as part of the functional tests which will fail as the gigadb-worker
couldn't be started beforehand.

Add full migrations run for FUW DB to the setup_devdb.sh which can (and should) be run
before both starting up the workers and the functional tests.