Add more contextual help to the file upload wizard flow and fix UX
the instructions and tips were added in the second step (metadata form)
to explain the bulk upload of metatada.

Update nginx conf to access the files directory so spreadsheet example can be linked to.

Temporarily disable XDEBUG on test container service as pecl/xdebug now requires
at et least PHP 7.2

Minor update to the instructions panel example in pattern library to match the usage
for contextual help.

in the breadcrumb, add link back to user's profile.