Correct destination of redirect for upload related dataset status
when the dataset is neither private/hidden or public, the redirect after updating
the dataset as an admin will lead to an error page saying the dataset cannot be shown.
Instead we redirect the admin to the dataset admin page.

We do this by adding a third dataset state (beside hidden and public) in DatasetPageSettings.php: draft

Updated unit tests and acceptance tests to cover that use case