Add console command to remove FUW drop-box
In the long run, this command will be part of the house keeping workflow.
But more immediately I need it to reset test data on cloud deployment so I can run
end to end test repeatedly.

Its purpose is to delete a dropbox created by FUW and associated data.
This cover the filesystem, the ftp accounts and the database tables (that we
don't really delete but change the status column as archived).
Finally the filedrop_account record in the FUW database has its status changed to STATUS_TERMINATED.

Basic functional test is added.