Fix the production version of config file that was missing a key
the download_base_url key of param-local.php was not present in its production conterpart,
so the file upload were reverting to the default (localhost).

Also the default url wasn't in a url format, so fix that.
Finally reordered the key in the dev config file to be consistent with production and not
getting mixed with ftp related key.